Mount the vue application to the parent of our top component
Vue 3 needs all vue components to be the descendants of
the element that we mount the application to.  In Vue 2,
we could mount to a component element.

This moves the mount point to the lux-library-header's parent
component, <body>, so that Vue can render the header.

See https://v3-migration.vuejs.org/breaking-changes/mount-changes.html